+++ src/tidypy/__init__.py
@@ -1,6 +1,4 @@
 
-# pylint: disable=wildcard-import
-
 from . import util
 
 from .collector import Collector
@@ -21,7 +19,11 @@ from .core import (
     execute_reports,
 )
 
-from .extenders import *
+from .extenders import (
+    Extender,
+    ExtenderError,
+    DoesNotExistError,
+)
 
 from .finder import Finder
 
@@ -31,7 +33,55 @@ from .progress import (
     ConsoleProgress,
 )
 
-from .reports import *
+from .reports import (
+    Report,
+)
 
-from .tools import *
+from .tools import (
+    Tool,
+    PythonTool,
+    Issue,
+    TidyPyIssue,
+    UnknownIssue,
+    AccessIssue,
+    ParseIssue,
+    ToolIssue,
+)
+
+
+__all__ = (
+    'execute_tools',
+    'execute_reports',
+    'get_tools',
+    'get_reports',
+    'get_extenders',
+    'get_default_config',
+    'get_user_config',
+    'get_local_config',
+    'get_project_config',
+    'purge_config_cache',
+    'util',
+
+    'Tool',
+    'PythonTool',
+    'Issue',
+    'TidyPyIssue',
+    'UnknownIssue',
+    'AccessIssue',
+    'ParseIssue',
+    'ToolIssue',
+
+    'Finder',
+    'Collector',
+
+    'Report',
+
+    'Extender',
+    'ExtenderError',
+    'DoesNotExistError',
+
+    'Progress',
+    'QuietProgress',
+    'ConsoleProgress',
+)